isgd.pl: Only shorten URLs on command
URLs are now no longer shortened automatically upon receiving them.
This prevents stuff from being shortened that you don't really want
to be sent to is.gd (maybe due to security reasons or whatever).
Instead we shorten the last found URL in the current buffer only
when the /isgd command is used.