pushover.pl 2.0
Support for other services has been removed in favor of the new
hsignal "pushover".

This can be used with any /trigger (for example combined with /exec)
or from other weechat scripts. That way it's very easy to implement
different notification services without having to reimplement the
entire "when to notify" logic and without bloating this script.